\name{identity_function}
\alias{identity_function}
\title{Function to map micro_process onto macro_model within calls to \code{AMME}.
}
\description{ A function to control for a node-level \code{micro_process} in \code{AMME} estimation.
}
\usage{
identity_function(x)
}
\arguments{
  \item{x}{a network object used to transfer \code{micro_process}.
}
}
\value{No return value, used internally with \code{AMME}}

