\name{Moran_dv}
\alias{Moran_dv}
\title{Function to calculate Moran's first order network autocorrelation in co-evolution SAOM.
}
\description{ A function to calculate Moran's first order network autocorrelation in co-evolution SAOM using both the endogenous dependent variables (behavior and network functions).
}
\usage{
Moran_dv(network)
}
\arguments{
  \item{network}{a network object used to calculate autocorrelation.
}
}
\value{No return value, used internally with \code{MEMS} and \code{AMME}}

